Converted toolbox position to i32 to make toolbox calculations easier

+++ src/component/static_grid_tile_prefab.rs
@@ -12,7 +12,7 @@ use super::{GhostColor, GhostColorComponent, GoalTile, GridPosition, JunctionTil
 #[derive(Debug, Serialize, Deserialize)]
 pub enum TilePosition {
     Grid { x: usize, y: usize },
-    Toolbox { toolbox: usize, index: usize },
+    Toolbox { toolbox: usize, index: i32 },
 }
 // TODO: follow the tutorial at [1]
 // TODO: add this to the level definition
